+++ NEWS.md
@@ -3,6 +3,8 @@ httpuv 1.4.2.9000
 
 * Fixed [#127](https://github.com/rstudio/httpuv/issues/127): Compilation failed on some platforms because `NULL` was used instead of an `Rcpp::List`. ([#131](https://github.com/rstudio/httpuv/pulls/131))
 
+* Fixed [#133](https://github.com/rstudio/httpuv/issues/133): Assertion failures when running on Fedora 28.
+
 httpuv 1.4.2
 ============